Kbuild: restore DWLAN_NL80211_TESTMODE flag
The DWLAN_NL80211_TESTMODE flag was removed on initial work of the driver at
commit 339af02c059112f684cff9447666bfbe99a0ea5d.
Bring it back again to be able to compile the driver when the kernel has
enabled the CONFIG_NL80211_TESTMODE option.

Signed-off-by: Isaac Hermida <isaac.hermida@digi.com>

Kbuild: add TESTMODE flag conditionally

QCA base code is adding TESTMODE flag inconditionally, but it requires
the kernel option CONFIG_NL80211_TESTMODE, that is enabled in Android default
kernel but not in DEY kernel.

Signed-off-by: Isaac Hermida <isaac.hermida@digi.com>