wlan: Update wlan config to provision p2p0 MAC Address
The Qualcomm WLAN driver supports several interfaces, like wlan0
and p2p0. It allows to use custom MAC addresses for each interface,
but the manufacturers can choose to provision a single MAC Address
that will be assigned to wlan0, and the MAC Address for p2p0 will
be derived from the wlan0 MAC Address by setting the "locally administered"
address bit. We provide a mechanism to use different MAC addresses
for each interface, then disable the "isP2pDeviceAddrAdministrated"
configuration item to notify the driver which provisioning is being
used.

https://jira.digi.com/browse/DEL-6612

Signed-off-by: Arturo Buzarra <arturo.buzarra@digi.com>