WCNSS_qcom_cfg: configure Strict Regulatory Mode
When Strict Regulatory Mode is enabled, the final frequency and power
restrictions are computed as an intersection of the ones from the programmed
regulatory domain and any dynamic domain configuration, whether it is performed
via 802.11d or manually by the user.

This way a domain configuration can only become equally or more restrictive
than the programmed regulatory domain.

https://jira.digi.com/browse/DEL-6819

Signed-off-by: Gonzalo Ruiz <Gonzalo.Ruiz@digi.com>