vos_api: use request_firmware_direct() to avoid boot delays due to optional fw
When the kernel's fw user helper is enabled and the driver requests firmware
files that don't exist in our filesystem, the uevent will remain active until
it times out (around 2 minutes). This causes the systemd-udev-settle service
to hang on boot until its own timeout expires.

Use request_firmware_direct() so the request fails right away if the firmware
doesn't exist in the fs, instead of waiting for someone to load it (which will
never happen).

https://jira.digi.com/browse/DEL-7114

Signed-off-by: Gabriel Valcazar <gabriel.valcazar@digi.com>