qcacld-2.0: Remove CFG80211_EXTERNAL_AUTH_SUPPORT define from SAE support
This define was used in combination with WLAN_FEATURE_SAE by the host
controller to advise to userspace about SAE support (NL80211_FEATURE_SAE flag).
However, kernel or user space do not provide this define, so by removing it, we
manage the SAE support only with the "sae_enabled" setting.

https://jira.digi.com/browse/DEL-7301

Signed-off-by: Arturo Buzarra <arturo.buzarra@digi.com>