qcacld: support building under DEA-11
As of Android 11, Google does not allow to use external tools (such as
make) inside Android's build system. NXP has implemented a set of makefiles
to build the bootloader, kernel, and modules externally before building
Android, and then use those artifacts as inputs for Android's build system.

Thus, this change removes the support for building the module in the
Android.mk makefile and implements a new makefile that is used to build
the module externally before building android.

https://jira.digi.com/browse/ADK4A-1803

Signed-off-by: Javier Viguera <javier.viguera@digi.com>