Respect and use the value for WLAN_ROOT consistently and pervasively.
The variable 'WLAN_ROOT' is intended to point at the root of the project
source tree and the location of Kbuild, which is what the build Linux
expects 'M' to be set to.

Consequently, Use the value for WLAN_ROOT consistently and pervasively
everywhere 'M' is set and passed to the Linux build.

In particular, 'PWD' will be incorrect if make is invoked with '-C'
and '$(shell pwd)' may be in conflict with a value of 'WLAN_ROOT'
asserted in the environment.