Use CURDIR rather than PWD as the default for WLAN_ROOT.
When make is invoked with '-C' from a directory other than the root
of the project source tree, the value of PWD will be the directory
where make was invoked. Consequently, it is only correct if and
only if make is invoked from the project source tree. However, the
make built-in variable CURDIR is always the directory of the current
makefile, regardless of where make was invoked from and the '-C'
option.